daemon,devicestate: refactor ModelAndGadgetInfo to SystemAndGadgetInfo()
